+++ InsertIntoCart.php
@@ -2,9 +2,9 @@
 
 
 
-require_once  __DIR__ .'/src/MovieRepository.php';
+require_once  __DIR__ .'/src/Movie.php';
 require_once  __DIR__ .'/src/Booking.php';
-$movieRepository = new MovieRepository ();
+$movieRepository = new Movie ();
 $movie= $movieRepository -> getIdByDateAndTime  ($_POST['id_film_'],$_POST['film_date_id_'],$_POST['time']);
